+++ verify/vite/vite.config.ts
@@ -1,0 +1,12 @@
+/* --------------------------------------------------------------------------------------------
+ * Copyright (c) 2024 TypeFox and others.
+ * Licensed under the MIT License. See LICENSE in the package root for license information.
+ * ------------------------------------------------------------------------------------------ */
+
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+    preview: {
+        port: 8082
+    }
+});